[♻️refactor]: api 호출 코드 리팩토링
